Gets ACS data from the Census API

Requests ACS population data for all tracts in the 50 states + DC as well as spatial data for the tracts. It also converts the spatial data to the given CRS.

fetch_tract_data(year, crs)

Arguments

year
Year of the ACS data and tract boundaries to fetch

crs
The EPSG code for the tract spatial data CRS

Value

A spatial data frame:

tract_fips: Census Tract FIPS code

year: Year

tract_pop: Census Tract population

tract_poverty_pop: Census Tract population below poverty line

tract_renting_hh: Number of renter households in Census Tract

geometry: Census Tract border spatial data

See also